Implement smart sprinklers
This implements the smart sprinklers algorithm with rain support.
Each circuit has now has a watering rate, and a goal per week, and we
track the trailing 2 weeks of watering cycles against our goal based on
the amount of water delivered.  If rain falls, the ammount that falls
will be recorded against all the circuits.  In the event of a large storm,
the circuits can delay as much as 2 weeks before resuming.